add validation

+++ app/src/main/java/com/example/crudapp/ShowPostActivity.java
@@ -1,20 +1,22 @@
- package com.example.crudapp;
+package com.example.crudapp;
 
- import android.content.Intent;
- import android.os.Bundle;
- import android.widget.Button;
- import android.widget.TextView;
+import android.content.Intent;
+import android.os.Bundle;
+import android.text.method.ScrollingMovementMethod;
+import android.widget.Button;
+import android.widget.TextView;
 
- import androidx.appcompat.app.AppCompatActivity;
- import androidx.lifecycle.ViewModelProvider;
+import androidx.appcompat.app.AlertDialog;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.lifecycle.ViewModelProvider;
 
- import com.example.crudapp.architecture.PostViewModel;
+import com.example.crudapp.architecture.PostViewModel;
 
- public class ShowPostActivity extends AppCompatActivity {
+public class ShowPostActivity extends AppCompatActivity {
 
     TextView postTitleView, postDescView, postAuthorView, postDateView, upvoteCountView, downvoteCountView;
     Button deleteBtn, editBtn;
-     private PostViewModel postViewModel;
+    private PostViewModel postViewModel;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -29,6 +31,8 @@
         downvoteCountView = findViewById(R.id.downvote_count_view);
         deleteBtn = findViewById(R.id.delete_btn);
         editBtn = findViewById(R.id.edit_btn);
+        postDescView.setMovementMethod(new ScrollingMovementMethod());
+        postTitleView.setMovementMethod(new ScrollingMovementMethod());
 
         postViewModel = new ViewModelProvider(this).get(PostViewModel.class);
 
@@ -36,41 +40,53 @@
         String title = intent.getStringExtra("POST_TITLE");
         String desc = intent.getStringExtra("POST_DESC");
         String author = intent.getStringExtra("POST_AUTHOR");
-        String date =  intent.getStringExtra("POST_DATE");
-        int upvotes = intent.getIntExtra("POST_UPVOTE_COUNT",0);
-        int downvotes = intent.getIntExtra("POST_DOWNVOTE_COUNT",0);
+        String date = intent.getStringExtra("POST_DATE");
+        int upvotes = intent.getIntExtra("POST_UPVOTE_COUNT", 0);
+        int downvotes = intent.getIntExtra("POST_DOWNVOTE_COUNT", 0);
 
         postTitleView.setText(title);
         postDescView.setText(desc);
         postAuthorView.setText(author);
         postDateView.setText(date);
-        upvoteCountView.setText(""+upvotes);
-        downvoteCountView.setText(""+downvotes);
+        upvoteCountView.setText("" + upvotes);
+        downvoteCountView.setText("" + downvotes);
 
-        int id = intent.getIntExtra("POST_ID",-1);
+        int id = intent.getIntExtra("POST_ID", -1);
 
         editBtn.setOnClickListener(view -> {
 
             Intent editIntent = new Intent(this, AddPostActivity.class);
-            editIntent.putExtra("ACTION_TYPE","EDIT");
-            editIntent.putExtra("POST_ID",id);
-            editIntent.putExtra("POST_TITLE",title);
-            editIntent.putExtra("POST_DESC",desc);
-            editIntent.putExtra("POST_AUTHOR",author);
-            editIntent.putExtra("POST_DATE",date);
-            editIntent.putExtra("POST_UPVOTE_COUNT",upvotes);
-            editIntent.putExtra("POST_DOWNVOTE_COUNT",downvotes);
+            editIntent.putExtra("ACTION_TYPE", "EDIT");
+            editIntent.putExtra("POST_ID", id);
+            editIntent.putExtra("POST_TITLE", title);
+            editIntent.putExtra("POST_DESC", desc);
+            editIntent.putExtra("POST_AUTHOR", author);
+            editIntent.putExtra("POST_DATE", date);
+            editIntent.putExtra("POST_UPVOTE_COUNT", upvotes);
+            editIntent.putExtra("POST_DOWNVOTE_COUNT", downvotes);
             startActivity(editIntent);
             this.finish();
 
         });
 
         deleteBtn.setOnClickListener(view -> {
-            PostModel post = new PostModel(title,desc,author,date,upvotes,downvotes);
-            post.setId(id);
-            postViewModel.delete(post);
-            startActivity(new Intent(this,MainActivity.class));
-            this.finish();
+
+            AlertDialog.Builder builder = new AlertDialog.Builder(this);
+            builder.setTitle("Confirm Deletion");
+            builder.setMessage("Are you sure you want to delete this post?");
+            builder.setPositiveButton("Delete", (dialog, which) -> {
+                PostModel post = new PostModel(title, desc, author, date, upvotes, downvotes);
+                post.setId(id);
+                postViewModel.delete(post);
+                startActivity(new Intent(this, MainActivity.class));
+                this.finish();
+            });
+            builder.setNegativeButton("Cancel", (dialog, which) -> {
+                // do nothing
+            });
+            AlertDialog dialog = builder.create();
+            dialog.show();
+
         });
 
     }
